requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pybit>=5.6.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
